// apps/farcaster-bot/__generated__/airstack/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Address: { input: any; output: any; }
  Any: { input: any; output: any; }
  DateRange: { input: any; output: any; }
  Identity: { input: any; output: any; }
  IntString: { input: any; output: any; }
  Map: { input: any; output: any; }
  Range: { input: any; output: any; }
  Time: { input: any; output: any; }
  TimeRange: { input: any; output: any; }
};
// ...
export enum Blockchain {
  Ethereum = 'ethereum'
}

export type Boolean_Comparator_Exp = {
  _eq?: InputMaybe<Scalars['Boolean']['input']>;
};

export type ConnectedAddress = {
  __typename?: 'ConnectedAddress';
  address?: Maybe<Scalars['Address']['output']>;
  blockchain?: Maybe<Scalars['String']['output']>;
  chainId?: Maybe<Scalars['String']['output']>;
  timestamp?: Maybe<Scalars['Time']['output']>;
};
// ...
export type Media = {
  __typename?: 'Media';
  animation_url?: Maybe<AnimationUrlVariants>;
  audio?: Maybe<AudioVariants>;
  image?: Maybe<ImageSizes>;
  json?: Maybe<Scalars['String']['output']>;
  video?: Maybe<VideoVariants>;
};
// ...
export type Query = {
  __typename?: 'Query';
  Accounts?: Maybe<AccountsOutput>;
  Domains?: Maybe<DomainsOutput>;
  FarcasterCasts?: Maybe<FarcasterCastOutput>;
  FarcasterChannelParticipants?: Maybe<FarcasterChannelParticipantsOutput>;
  FarcasterChannels?: Maybe<FarcasterChannelsOutput>;
  FarcasterQuotedRecasts?: Maybe<FarcasterQuotedRecastsOutput>;
  FarcasterReactions?: Maybe<FarcasterReactionsOutput>;
  FarcasterReplies?: Maybe<FarcasterRepliesOutput>;
  FarcasterValidateFrameMessage?: Maybe<FarcasterFrameMessageOutput>;
  PoapEvents?: Maybe<PoapEventsOutput>;
  Poaps?: Maybe<PoapsOutput>;
  Snapshots?: Maybe<SnapshotsOutput>;
  SocialFollowers?: Maybe<SocialFollowerOutput>;
  SocialFollowings?: Maybe<SocialFollowingOutput>;
  Socials?: Maybe<SocialsOutput>;
  TokenBalances?: Maybe<TokenBalancesOutput>;
  TokenNfts?: Maybe<TokenNftsOutput>;
  TokenTransfers?: Maybe<TokenTransfersOutput>;
  Tokens?: Maybe<TokensOutput>;
  TrendingCasts?: Maybe<TrendingCastsOutput>;
  TrendingMints?: Maybe<TrendingMintsOutput>;
  TrendingSwaps?: Maybe<TrendingSwapsOutput>;
  TrendingTokens?: Maybe<TrendingTokensOutput>;
  Wallet?: Maybe<Wallet>;
  XMTPs?: Maybe<XmtPsOutput>;
};
// ...
export type Social = {
  __typename?: 'Social';
  /** Blockchain associated with the social identity */
  blockchain?: Maybe<Blockchain>;
  /** Unique identifier for the blockchain */
  chainId?: Maybe<Scalars['String']['output']>;
  connectedAddresses?: Maybe<Array<ConnectedAddress>>;
  coverImageContentValue?: Maybe<Media>;
  coverImageURI?: Maybe<Scalars['String']['output']>;
  /** Social DApp name */
  dappName?: Maybe<SocialDappName>;
  /** Social DApp slug (contract version) */
  dappSlug?: Maybe<SocialDappSlug>;
  /** Airstack unique dapp version number */
  dappVersion?: Maybe<Scalars['String']['output']>;
  fnames?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  followerCount?: Maybe<Scalars['Int']['output']>;
  followerTokenAddress?: Maybe<Scalars['Address']['output']>;
  followers?: Maybe<SocialFollowerOutput>;
  followingCount?: Maybe<Scalars['Int']['output']>;
  followings?: Maybe<SocialFollowingOutput>;
  handleTokenAddress?: Maybe<Scalars['Address']['output']>;
  handleTokenId?: Maybe<Scalars['String']['output']>;
  /** Airstack unique identifier for the data point */
  id?: Maybe<Scalars['ID']['output']>;
  /** Blockchain address, ENS domain name, social identity such as Farcaster (for Farcaster use 'fc_fid:' prefix followed by the Farcaster user ID like fc_fid:5650, or use 'fc_fname:' prefix followed by the Farcaster user ID like 'fc_fname:vbuterin') or Lens (e.g. 'stani.lens) */
  identity?: Maybe<Scalars['Identity']['output']>;
  isDefault?: Maybe<Scalars['Boolean']['output']>;
  isFarcasterPowerUser?: Maybe<Scalars['Boolean']['output']>;
  location?: Maybe<Scalars['String']['output']>;
  metadataURI?: Maybe<Scalars['String']['output']>;
  profileBio?: Maybe<Scalars['String']['output']>;
  profileCreatedAtBlockNumber?: Maybe<Scalars['Int']['output']>;
  profileCreatedAtBlockTimestamp?: Maybe<Scalars['Time']['output']>;
  profileDisplayName?: Maybe<Scalars['String']['output']>;
  profileHandle?: Maybe<Scalars['String']['output']>;
  profileImage?: Maybe<Scalars['String']['output']>;
  profileImageContentValue?: Maybe<Media>;
  profileLastUpdatedAtBlockNumber?: Maybe<Scalars['Int']['output']>;
  profileLastUpdatedAtBlockTimestamp?: Maybe<Scalars['Time']['output']>;
  profileMetadata?: Maybe<Scalars['Map']['output']>;
  profileName?: Maybe<Scalars['String']['output']>;
  profileTokenAddress?: Maybe<Scalars['String']['output']>;
  profileTokenId?: Maybe<Scalars['String']['output']>;
  profileTokenIdHex?: Maybe<Scalars['String']['output']>;
  profileTokenUri?: Maybe<Scalars['String']['output']>;
  profileUrl?: Maybe<Scalars['String']['output']>;
  socialCapital?: Maybe<SocialCapital>;
  twitterUserName?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['Time']['output']>;
  userAddress?: Maybe<Scalars['Address']['output']>;
  userAddressDetails?: Maybe<Wallet>;
  userAssociatedAddressDetails?: Maybe<Array<Wallet>>;
  /** blockchain addresses associated with the social profile */
  userAssociatedAddresses?: Maybe<Array<Scalars['Address']['output']>>;
  userCreatedAtBlockNumber?: Maybe<Scalars['Int']['output']>;
  userCreatedAtBlockTimestamp?: Maybe<Scalars['Time']['output']>;
  userHomeURL?: Maybe<Scalars['String']['output']>;
  userId?: Maybe<Scalars['String']['output']>;
  userLastUpdatedAtBlockNumber?: Maybe<Scalars['Int']['output']>;
  userLastUpdatedAtBlockTimestamp?: Maybe<Scalars['Time']['output']>;
  userRecoveryAddress?: Maybe<Scalars['Address']['output']>;
  website?: Maybe<Scalars['String']['output']>;
};


export type SocialFollowersArgs = {
  input?: InputMaybe<SocialFollowerNestedInput>;
};


export type SocialFollowingsArgs = {
  input?: InputMaybe<SocialFollowingNestedInput>;
};

export type SocialCapital = {
  __typename?: 'SocialCapital';
  socialCapitalRank?: Maybe<Scalars['Int']['output']>;
  socialCapitalScore?: Maybe<Scalars['Float']['output']>;
  socialCapitalScoreRaw?: Maybe<Scalars['String']['output']>;
};

export type SocialCapitalValue = {
  __typename?: 'SocialCapitalValue';
  formattedValue?: Maybe<Scalars['Float']['output']>;
  hash?: Maybe<Scalars['String']['output']>;
  rawValue?: Maybe<Scalars['String']['output']>;
};

export enum SocialDappName {
  Farcaster = 'farcaster',
  Lens = 'lens'
}

export type SocialDappName_Comparator_Exp = {
  _eq?: InputMaybe<SocialDappName>;
  _in?: InputMaybe<Array<SocialDappName>>;
};

export enum SocialDappSlug {
  FarcasterGoerli = 'farcaster_goerli',
  FarcasterOptimism = 'farcaster_optimism',
  FarcasterV2Optimism = 'farcaster_v2_optimism',
  FarcasterV3Optimism = 'farcaster_v3_optimism',
  LensPolygon = 'lens_polygon',
  LensV2Polygon = 'lens_v2_polygon'
}
// ...
export type SocialFollowerOutput = {
  __typename?: 'SocialFollowerOutput';
  Follower?: Maybe<Array<SocialFollower>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type SocialFollowingOutput = {
  __typename?: 'SocialFollowingOutput';
  Following?: Maybe<Array<SocialFollowing>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type SocialsOutput = {
  __typename?: 'SocialsOutput';
  Social?: Maybe<Array<Social>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type Wallet = {
  __typename?: 'Wallet';
  /** Represent On-chain smart-contract accounts */
  accounts?: Maybe<Array<Account>>;
  /** Returns addresses associated with the identity input */
  addresses?: Maybe<Array<Scalars['Address']['output']>>;
  /** Blockchain associated with the provided identity */
  blockchain?: Maybe<TokenBlockchain>;
  /** Nested query - allows querying domains owned by the address */
  domains?: Maybe<Array<Domain>>;
  /** Blockchain address, ENS domain name, social identity such as Farcaster (for Farcaster use 'fc_fid:' prefix followed by the Farcaster user ID like fc_fid:5650, or use 'fc_fname:' prefix followed by the Farcaster user ID like 'fc_fname:vbuterin') or Lens (e.g. 'stani.lens) */
  identity: Scalars['Identity']['output'];
  /** Returns Poaps owned by the address */
  poaps?: Maybe<Array<Poap>>;
  /** Nested query - allows returning primary domains, if applicable */
  primaryDomain?: Maybe<Domain>;
  socialFollowers?: Maybe<SocialFollowerOutput>;
  socialFollowings?: Maybe<SocialFollowingOutput>;
  /** Returns social profile information related to the address */
  socials?: Maybe<Array<Social>>;
  /** Nested query - allows returning token balances */
  tokenBalances?: Maybe<Array<TokenBalance>>;
  /** Nested query - allows returning token transfers and related information */
  tokenTransfers?: Maybe<Array<TokenTransfer>>;
  /** Nested query - allows querying the XMTP enabled addresses */
  xmtp?: Maybe<Array<Xmtp>>;
};
// ...
export type GetFarcasterSocialsByAddressQueryVariables = Exact<{
  identity: Scalars['Identity']['input'];
}>;


export type GetFarcasterSocialsByAddressQuery = { __typename?: 'Query', Socials?: { __typename?: 'SocialsOutput', Social?: Array<{ __typename?: 'Social', profileHandle?: string | null, profileTokenId?: string | null }> | null } | null };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}

export const GetFarcasterSocialsByAddressDocument = new TypedDocumentString(`
    query GetFarcasterSocialsByAddress($identity: Identity!) {
  Socials(
    input: {filter: {identity: {_eq: $identity}, dappName: {_eq: farcaster}}, blockchain: ethereum, order: {followerCount: DESC}, limit: 1}
  ) {
    Social {
      profileHandle
      profileTokenId
    }
  }
}
    `) as unknown as TypedDocumentString<GetFarcasterSocialsByAddressQuery, GetFarcasterSocialsByAddressQueryVariables>;